Change ints to uint64_ts in stamp creator

+++ src/kbmod/search/stamp_creator.cpp
@@ -79,10 +79,10 @@ std::vector<RawImage> StampCreator::get_coadded_stamps_cpu(ImageStack& stack,
                                                            std::vector<Trajectory>& t_array,
                                                            std::vector<std::vector<bool>>& use_index_vect,
                                                            const StampParameters& params) {
-    const int num_trajectories = t_array.size();
+    const uint64_t num_trajectories = t_array.size();
     std::vector<RawImage> results(num_trajectories);
 
-    for (int i = 0; i < num_trajectories; ++i) {
+    for (uint64_t i = 0; i < num_trajectories; ++i) {
         std::vector<RawImage> stamps =
                 StampCreator::create_stamps(stack, t_array[i], params.radius, true, use_index_vect[i]);
 
@@ -161,17 +161,17 @@ std::vector<RawImage> StampCreator::get_coadded_stamps_gpu(ImageStack& stack,
         throw std::runtime_error("Invalid Radius.");
     }
 
-    const int num_images = stack.img_count();
-    const int width = stack.get_width();
-    const int height = stack.get_height();
+    const uint64_t num_images = stack.img_count();
+    const uint64_t width = stack.get_width();
+    const uint64_t height = stack.get_height();
 
     // Create a data stucture for the per-image data.
     std::vector<double> image_times = stack.build_zeroed_times();
 
     // Allocate space for the results.
-    const int num_trajectories = t_array.size();
-    const int stamp_width = 2 * params.radius + 1;
-    const int stamp_ppi = stamp_width * stamp_width;
+    const uint64_t num_trajectories = t_array.size();
+    const uint64_t stamp_width = 2 * params.radius + 1;
+    const uint64_t stamp_ppi = stamp_width * stamp_width;
     std::vector<float> stamp_data(stamp_ppi * num_trajectories);
 
     // Do the co-adds.
@@ -200,9 +200,9 @@ std::vector<RawImage> StampCreator::get_coadded_stamps_gpu(ImageStack& stack,
     // Copy the stamps into RawImages and do the filtering.
     std::vector<RawImage> results(num_trajectories);
     std::vector<float> current_pixels(stamp_ppi, 0.0);
-    for (int t = 0; t < num_trajectories; ++t) {
+    for (uint64_t t = 0; t < num_trajectories; ++t) {
         // Copy the data into a single RawImage.
-        int offset = t * stamp_ppi;
+        uint64_t offset = t * stamp_ppi;
         for (unsigned p = 0; p < stamp_ppi; ++p) {
             current_pixels[p] = stamp_data[offset + p];
         }
